# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Band service API in Go",
        "title": "Band Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/bands": {
            "get": {
                "description": "returns a list of all bands",
                "tags": [
                    "Bands"
                ],
                "summary": "returns a list of all bands",
                "responses": {}
            },
            "post": {
                "description": "creates a band",
                "tags": [
                    "Bands"
                ],
                "summary": "creates a band",
                "parameters": [
                    {
                        "description": "Create band",
                        "name": "band",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateBandRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/bands/{bandId}": {
            "get": {
                "description": "returns a band by its id",
                "tags": [
                    "Bands"
                ],
                "summary": "returns a band by its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "find band by id",
                        "name": "bandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "deletes a band by its id",
                "tags": [
                    "Bands"
                ],
                "summary": "deletes a band by its id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "find band by id",
                        "name": "bandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "updates a band",
                "tags": [
                    "Bands"
                ],
                "summary": "updates a band",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update band by id",
                        "name": "bandId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update band",
                        "name": "band",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateBandRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "request.CreateBandRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        }
    }
}